Initial prototype
This prototype is intended to replace an older PHP script
used to query the Meta Microsoft SQL Server database
and generate a pipe-delimited input file containing
patron records for further processing.

While testing shows that feature parity is provided, further
work is needed.

This code should be considered to be in a "beta" state. Testing
indicates that it is functional, but some edge cases likely
remain and further enhancements are needed to "round off"
the edges.

refs help-157421