Move stubbed output file settings to Config struct
These settings are not specific to fileConfig or cliConfig
nested structs, but apply equally to both.

Not sure if/when support will be added for specifying
output files, but they're more relevant in the new
location. Moving now to help make clear the intent that
all config file-specific settings are to be excluded from
flag parsing.

refs GH-23